uniswapv3: all tests passing for swap router and supporting functions
